# openapi.json
{"openapi":"3.1.0","info":{"title":"QEW Innovation Corridor API","description":"Unified backend for camera scraping and AI direction analysis","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","description":"Root endpoint","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/scraper/start":{"post":{"summary":"Start Scraper","description":"Start camera scraping process","operationId":"start_scraper_api_scraper_start_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScraperStartRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/scraper/stop":{"post":{"summary":"Stop Scraper","description":"Stop camera scraping process","operationId":"stop_scraper_api_scraper_stop_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/scraper/status":{"get":{"summary":"Get Scraper Status","description":"Get camera scraper status","operationId":"get_scraper_status_api_scraper_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/cameras/latest":{"get":{"summary":"Get Latest Cameras","description":"Get latest camera data with most recent images","operationId":"get_latest_cameras_api_cameras_latest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/cameras/{camera_id}":{"get":{"summary":"Get Camera","description":"Get specific camera by ID","operationId":"get_camera_api_cameras__camera_id__get","parameters":[{"name":"camera_id","in":"path","required":true,"schema":{"type":"integer","title":"Camera Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/collections/latest":{"get":{"summary":"Get Latest Collection","description":"Get latest collection info","operationId":"get_latest_collection_api_collections_latest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/stats":{"get":{"summary":"Get Stats","description":"Get database statistics","operationId":"get_stats_api_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/directions/analyze":{"post":{"summary":"Analyze Direction","description":"Analyze direction for a specific camera view","operationId":"analyze_direction_api_directions_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirectionAnalysisRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/directions/analyze-batch":{"post":{"summary":"Analyze Batch","description":"Analyze all pending cameras","operationId":"analyze_batch_api_directions_analyze_batch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchAnalysisRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/directions/stop":{"post":{"summary":"Stop Direction Analysis","description":"Stop direction analysis process","operationId":"stop_direction_analysis_api_directions_stop_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/directions/status":{"get":{"summary":"Get Direction Status","description":"Get direction analysis status","operationId":"get_direction_status_api_directions_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/directions/{camera_id}/{view_id}":{"get":{"summary":"Get Direction","description":"Get direction assessment for specific camera view","operationId":"get_direction_api_directions__camera_id___view_id__get","parameters":[{"name":"camera_id","in":"path","required":true,"schema":{"type":"integer","title":"Camera Id"}},{"name":"view_id","in":"path","required":true,"schema":{"type":"integer","title":"View Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/directions/pending":{"get":{"summary":"Get Pending Cameras","description":"Get list of cameras that need direction analysis","operationId":"get_pending_cameras_api_directions_pending_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/directions":{"get":{"summary":"Get All Directions","description":"Get all direction assessments","operationId":"get_all_directions_api_directions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/images/{collection_id}/{filename}":{"get":{"summary":"Get Image","description":"Serve camera image file","operationId":"get_image_api_images__collection_id___filename__get","parameters":[{"name":"collection_id","in":"path","required":true,"schema":{"type":"string","title":"Collection Id"}},{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BatchAnalysisRequest":{"properties":{"platform":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Platform","default":"Gemini"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model","default":"Gemini 2.0 Flash"}},"type":"object","title":"BatchAnalysisRequest"},"DirectionAnalysisRequest":{"properties":{"camera_id":{"type":"integer","title":"Camera Id"},"view_id":{"type":"integer","title":"View Id"},"platform":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Platform","default":"Gemini"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model","default":"Gemini 2.0 Flash"}},"type":"object","required":["camera_id","view_id"],"title":"DirectionAnalysisRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ScraperStartRequest":{"properties":{"images_per_camera":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Images Per Camera","default":1},"delay_between_captures":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Delay Between Captures","default":60}},"type":"object","title":"ScraperStartRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}